rest_api_calculator/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
